:wrench: isolate item_repository from item
